Punch given push ability
-_attempt_push and _attempt_knockback added to combat arena. They need
refactor, but are accomplishing their goal at the moment. attempt_push
is a slightly more generic function, and attempt_knockback is a cascade
effect. The reason for the separation was a necessity to limit chaining
effects. Push also has been made in a way to allow for future
functionality, but again: a refactor is in order and thus might be
possible to make it even more functional for different circumstances.