Refactor of character, basic shot, and heal
-Changed Character and input process to use a default ability (referred
to as BASE_ATTACK) instead of manually programmed action. The base shot
has been created for this purpose, as test-shot. And both red and blue
character have been updated for this.

-linear_search has been replaced with a sequence of helper functions
that will allow nesting of row, column, and/or specifying character vs
tile, with extensibility for future search parameters to be added as a
dictionary.

-capture_tile updated to use new search functions

-heal updated, to process through the arena (GameController).
_attempt_heal function includes an overheal boolean to allow for special
healing above max, in the future.

-HpNode changed to include a MAX_HP and local HP